+++ image/user/db/notes_001.sql
@@ -1,6 +1,5 @@
 CREATE TABLE IF NOT EXISTS notes(
     _id INTEGER PRIMARY KEY,
     date INTEGER,
-    snippet TEXT DEFAULT '',
-    path TEXT DEFAULT ''
+    snippet TEXT DEFAULT ''
 );