+++ module-apps/application-notes/model/NotesListModel.hpp
@@ -3,35 +3,41 @@
 
 #pragma once
 
+#include <cstdint>
 #include <vector>
 
-#include "Interface/NotesRecord.hpp"
-#include "DatabaseModel.hpp"
-#include "NotesRecord.hpp"
-#include "Application.hpp"
-#include "ListItemProvider.hpp"
+#include <module-apps/Application.hpp>
+#include <module-apps/DatabaseModel.hpp>
+
+#include <module-db/Interface/NotesRecord.hpp>
+
+#include <module-gui/gui/widgets/ListItemProvider.hpp>
+
+#include "NotesRepository.hpp"
 
 namespace app::notes
 {
-    class NotesProvider : public app::DatabaseModel<NotesRecord>, public gui::ListItemProvider
+    class NotesListItemProvider : public app::DatabaseModel<NotesRecord>, public gui::ListItemProvider
     {
       public:
-        explicit NotesProvider(app::Application *app);
+        explicit NotesListItemProvider(app::Application *app);
     };
 
-    class NotesListModel : public NotesProvider
+    class NotesListModel : public NotesListItemProvider
     {
       public:
-        explicit NotesListModel(app::Application *app);
+        NotesListModel(app::Application *app, std::shared_ptr<AbstractNotesRepository> notesRepository);
+
+        [[nodiscard]] unsigned int requestRecordsCount() override;
+        [[nodiscard]] bool updateRecords(std::vector<NotesRecord> records) override;
+        void requestRecords(std::uint32_t offset, std::uint32_t limit) override;
 
         [[nodiscard]] gui::ListItem *getItem(gui::Order order) override;
         [[nodiscard]] unsigned int getMinimalItemHeight() const override;
 
-        unsigned int requestRecordsCount() override;
-        bool updateRecords(std::vector<NotesRecord> records) override;
-        void requestRecords(const uint32_t offset, const uint32_t limit) override;
-
       private:
-        bool handleQueryResponse(db::QueryResult *queryResult);
+        bool onNotesRetrieved(const std::vector<NotesRecord> &records, unsigned int notesRepoCount);
+
+        std::shared_ptr<AbstractNotesRepository> notesRepository;
     };
 } // namespace app::notes