+++ module-apps/application-notes/widgets/NotesItem.cpp
@@ -6,9 +6,16 @@
 #include <Style.hpp>
 #include <module-apps/application-notes/style/NotesListStyle.hpp>
 
+#include <module-utils/time/time_conversion.hpp>
+
 namespace gui
 {
-    NotesItem::NotesItem(std::shared_ptr<NotesRecord> record, bool mode24H) : note{std::move(record)}, mode24H{mode24H}
+    NotesItem::NotesItem(std::shared_ptr<NotesRecord> record) : note{std::move(record)}
+    {
+        buildInterface();
+    }
+
+    void NotesItem::buildInterface()
     {
         namespace notesItemStyle = app::notes::style::list::item;
         setMinimumSize(style::window::default_body_width, notesItemStyle::Height);
@@ -26,9 +33,14 @@ namespace gui
         date->setFont(style::window::font::medium);
         date->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Right, gui::Alignment::Vertical::Top});
 
-        snippet = createEmptyLabel(this);
+        snippet = new gui::TextFixedSize(this, 0, 0, 0, 0);
         snippet->setFont(style::window::font::small);
         snippet->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Left});
+        snippet->setEdges(gui::RectangleEdge::None);
+        snippet->setPenFocusWidth(::style::window::default_border_focus_w);
+        snippet->setPenWidth(::style::window::default_border_rect_no_focus);
+        snippet->setEditMode(gui::EditMode::BROWSE);
+        snippet->setUnderline(false);
 
         setSnippet(note->snippet);
         setDateText(note->date);
@@ -50,7 +62,12 @@ namespace gui
 
     void NotesItem::setDateText(std::uint32_t timestamp)
     {
-        date->setText(utils::to_string(timestamp));
+        if (auto dt = utils::time::DateTime(timestamp); dt.isYesterday()) {
+            date->setText(utils::localize.get("common_yesterday"));
+        }
+        else {
+            date->setText(dt);
+        }
     }
 
     bool NotesItem::onDimensionChanged(const BoundingBox &oldDim, const BoundingBox &newDim)
@@ -71,4 +88,4 @@ namespace gui
                              notesItemStyle::snippet::TopMargin);
         return true;
     }
-} /* namespace gui */
+} // namespace gui