+++ module-db/Interface/NotesRecord.cpp
@@ -3,23 +3,23 @@
 
 #include "NotesRecord.hpp"
 
-#include "queries/notes/NotesGetForList.hpp"
+#include "queries/notes/QueryNotesGet.hpp"
+#include "queries/notes/QueryNoteStore.hpp"
+#include "queries/notes/QueryNoteRemove.hpp"
 
 #include <cassert>
 
 namespace
 {
-    NotesTableFields toNotesTableFields(NotesRecordField field) noexcept
+    NotesTableFields toNotesTableFields(NotesRecordField field)
     {
         switch (field) {
         case NotesRecordField::Date:
             return NotesTableFields::Date;
         case NotesRecordField::Snippet:
             return NotesTableFields::Snippet;
-        case NotesRecordField::Path:
-            return NotesTableFields::Path;
         }
-        assert(false);
+        throw std::invalid_argument{"Invalid notes record field passed."};
     }
 } // namespace
 
@@ -28,29 +28,30 @@ NotesRecordInterface::NotesRecordInterface(NotesDB *notesDb) : notesDB(notesDb)
 
 bool NotesRecordInterface::Add(const NotesRecord &rec)
 {
-    return notesDB->notes.add(NotesTableRow{.date = rec.date, .snippet = rec.snippet, .path = rec.path.string()});
+    return notesDB->notes.add(NotesTableRow{.date = rec.date, .snippet = rec.snippet});
 }
 
-uint32_t NotesRecordInterface::GetCount()
+std::uint32_t NotesRecordInterface::GetCount()
 {
     return notesDB->notes.count();
 }
 
-std::unique_ptr<std::vector<NotesRecord>> NotesRecordInterface::GetLimitOffsetByField(uint32_t offset,
-                                                                                      uint32_t limit,
+std::unique_ptr<std::vector<NotesRecord>> NotesRecordInterface::GetLimitOffsetByField(std::uint32_t offset,
+                                                                                      std::uint32_t limit,
                                                                                       NotesRecordField field,
                                                                                       const char *str)
 {
     auto records     = std::make_unique<std::vector<NotesRecord>>();
     const auto notes = notesDB->notes.getLimitOffsetByField(offset, limit, toNotesTableFields(field), str);
     for (const auto &w : notes) {
-        NotesRecord record{w.ID, w.date, w.snippet, std::filesystem::path{w.path}};
+        NotesRecord record{w.ID, w.date, w.snippet};
         records->push_back(std::move(record));
     }
     return records;
 }
 
-std::unique_ptr<std::vector<NotesRecord>> NotesRecordInterface::GetLimitOffset(uint32_t offset, uint32_t limit)
+std::unique_ptr<std::vector<NotesRecord>> NotesRecordInterface::GetLimitOffset(std::uint32_t offset,
+                                                                               std::uint32_t limit)
 {
     return std::make_unique<std::vector<NotesRecord>>(getNotes(offset, limit));
 }
@@ -62,12 +63,16 @@ bool NotesRecordInterface::Update(const NotesRecord &rec)
         return false;
     }
 
-    notesDB->notes.update(
-        NotesTableRow{.ID = rec.ID, .date = rec.date, .snippet = rec.snippet, .path = rec.path.string()});
+    notesDB->notes.update(NotesTableRow{.ID = rec.ID, .date = rec.date, .snippet = rec.snippet});
     return true;
 }
 
-bool NotesRecordInterface::RemoveByID(uint32_t id)
+bool NotesRecordInterface::RemoveAll()
+{
+    return notesDB->notes.removeAll();
+}
+
+bool NotesRecordInterface::RemoveByID(std::uint32_t id)
 {
     return notesDB->notes.removeById(id);
 }
@@ -82,18 +87,18 @@ bool NotesRecordInterface::RemoveByField(NotesRecordField field, const char *str
     }
 }
 
-NotesRecord NotesRecordInterface::GetByID(uint32_t id)
+NotesRecord NotesRecordInterface::GetByID(std::uint32_t id)
 {
     auto note = notesDB->notes.getById(id);
-    return NotesRecord{note.ID, note.date, note.snippet, std::filesystem::path{note.path}};
+    return NotesRecord{note.ID, note.date, note.snippet};
 }
 
-std::vector<NotesRecord> NotesRecordInterface::getNotes(uint32_t offset, uint32_t limit) const
+std::vector<NotesRecord> NotesRecordInterface::getNotes(std::uint32_t offset, std::uint32_t limit) const
 {
     std::vector<NotesRecord> records;
     const auto notes = notesDB->notes.getLimitOffset(offset, limit);
     for (const auto &w : notes) {
-        NotesRecord record{w.ID, w.date, w.snippet, std::filesystem::path{w.path}};
+        NotesRecord record{w.ID, w.date, w.snippet};
         records.push_back(std::move(record));
     }
     return records;
@@ -101,18 +106,49 @@ std::vector<NotesRecord> NotesRecordInterface::getNotes(uint32_t offset, uint32_
 
 std::unique_ptr<db::QueryResult> NotesRecordInterface::runQuery(std::shared_ptr<db::Query> query)
 {
-    LOG_ERROR("NOTES COUNT: %u", GetCount());
-    if (typeid(*query) == typeid(db::query::NotesGetForList)) {
-        return getForListQuery(query);
+    if (typeid(*query) == typeid(db::query::QueryNotesGet)) {
+        return getQuery(query);
+    }
+    if (typeid(*query) == typeid(db::query::QueryNoteStore)) {
+        return storeQuery(query);
+    }
+    if (typeid(*query) == typeid(db::query::QueryNoteRemove)) {
+        return removeQuery(query);
     }
     return nullptr;
 }
 
-std::unique_ptr<db::QueryResult> NotesRecordInterface::getForListQuery(const std::shared_ptr<db::Query> &query)
+std::unique_ptr<db::QueryResult> NotesRecordInterface::getQuery(const std::shared_ptr<db::Query> &query)
 {
-    const auto localQuery = static_cast<db::query::NotesGetForList *>(query.get());
-    const auto &records   = getNotes(localQuery->offset, localQuery->limit);
-    auto response         = std::make_unique<db::query::NotesGetForListResult>(records);
+    const auto localQuery = static_cast<db::query::QueryNotesGet *>(query.get());
+    const auto &records   = getNotes(localQuery->getOffset(), localQuery->getLimit());
+    auto response         = std::make_unique<db::query::NotesGetResult>(records, GetCount());
+    response->setRequestQuery(query);
+    return response;
+}
+
+std::unique_ptr<db::QueryResult> NotesRecordInterface::storeQuery(const std::shared_ptr<db::Query> &query)
+{
+    const auto localQuery = static_cast<db::query::QueryNoteStore *>(query.get());
+    const auto &record    = localQuery->getRecord();
+    bool isSuccess        = false;
+    if (const auto exists = notesDB->notes.getById(record.ID).ID != DB_ID_NONE; exists) {
+        isSuccess = Update(record);
+    }
+    else {
+        isSuccess = Add(record);
+    }
+
+    auto response = std::make_unique<db::query::NoteStoreResult>(isSuccess);
+    response->setRequestQuery(query);
+    return response;
+}
+
+std::unique_ptr<db::QueryResult> NotesRecordInterface::removeQuery(const std::shared_ptr<db::Query> &query)
+{
+    const auto localQuery = static_cast<db::query::QueryNoteRemove *>(query.get());
+    const auto isSuccess  = RemoveByID(localQuery->getRecordId());
+    auto response         = std::make_unique<db::query::NoteRemoveResult>(isSuccess);
     response->setRequestQuery(query);
     return response;
 }