+++ module-db/Interface/NotesRecord.hpp
@@ -17,14 +17,12 @@ struct NotesRecord : public Record
 {
     std::uint32_t date;
     UTF8 snippet;
-    std::filesystem::path path;
 };
 
 enum class NotesRecordField
 {
     Date,
     Snippet,
-    Path
 };
 
 class NotesRecordInterface : public RecordInterface<NotesRecord, NotesRecordField>
@@ -33,23 +31,26 @@ class NotesRecordInterface : public RecordInterface<NotesRecord, NotesRecordFiel
     explicit NotesRecordInterface(NotesDB *notesDb);
 
     bool Add(const NotesRecord &rec) override;
-    bool RemoveByID(uint32_t id) override;
+    bool RemoveAll() override;
+    bool RemoveByID(std::uint32_t id) override;
     bool RemoveByField(NotesRecordField field, const char *str) override;
     bool Update(const NotesRecord &rec) override;
-    NotesRecord GetByID(uint32_t id) override;
-    uint32_t GetCount() override;
-    std::unique_ptr<std::vector<NotesRecord>> GetLimitOffset(uint32_t offset, uint32_t limit) override;
-    std::unique_ptr<std::vector<NotesRecord>> GetLimitOffsetByField(uint32_t offset,
-                                                                    uint32_t limit,
+    NotesRecord GetByID(std::uint32_t id) override;
+    std::uint32_t GetCount() override;
+    std::unique_ptr<std::vector<NotesRecord>> GetLimitOffset(std::uint32_t offset, std::uint32_t limit) override;
+    std::unique_ptr<std::vector<NotesRecord>> GetLimitOffsetByField(std::uint32_t offset,
+                                                                    std::uint32_t limit,
                                                                     NotesRecordField field,
                                                                     const char *str) override;
 
     std::unique_ptr<db::QueryResult> runQuery(std::shared_ptr<db::Query> query) override;
 
   private:
-    std::vector<NotesRecord> getNotes(uint32_t offset, uint32_t limit) const;
+    std::vector<NotesRecord> getNotes(std::uint32_t offset, std::uint32_t limit) const;
 
-    std::unique_ptr<db::QueryResult> getForListQuery(const std::shared_ptr<db::Query> &query);
+    std::unique_ptr<db::QueryResult> getQuery(const std::shared_ptr<db::Query> &query);
+    std::unique_ptr<db::QueryResult> storeQuery(const std::shared_ptr<db::Query> &query);
+    std::unique_ptr<db::QueryResult> removeQuery(const std::shared_ptr<db::Query> &query);
 
     NotesDB *notesDB;
 };